+++ validation/mounts.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/opencontainers/runtime-tools/validation/util"
+)
+
+func main() {
+	util.Skip("TODO: mounts generation options have not been implemented", "")
+}
